split app and topNavigation
