+++ src/components/App.jsx
@@ -11,20 +11,11 @@ import LandingPageTeaser from '../containers/LandingPageTeaser';
 import LandingPageContent from '../containers/LandingPageContent';
 import Dashboard from '../containers/Dashboard';
 import Login from '../containers/Login';
+import TopNavigation from '../containers/TopNavigation';
 
 import '../assets/landing-page.css';
 
 class App extends Component {
-  constructor(props) {
-    super(props);
-    this.gotoHomepage = this.gotoHomepage.bind(this);
-    this.onScreenShow = this.onScreenShow.bind(this);
-  }
-
-  gotoHomepage() {
-    this.props.changePage('landingpage');
-  }
-
   getPageContent() {
     switch (this.props.currentPage) {
       case 'learning':
@@ -77,18 +68,10 @@ class App extends Component {
 
   render() {
     const pageContent = this.getPageContent();
-    const loginButton = this.getLoginButton();
-    const dashboardButton = this.getDashboardButton();
 
     return (
       <div>
-        <nav className="navbar navbar-light bg-light static-top">
-          <div className="container">
-            <a className="navbar-brand" href="#" onClick={event=>this.gotoHomepage()}>kovoel</a>
-            {loginButton}
-            {dashboardButton}
-          </div>
-        </nav>
+        <TopNavigation />
 
         {pageContent}
 
@@ -98,42 +81,6 @@ class App extends Component {
 
       </div>
     );
-  }
-
-  getDashboardButton() {
-    let dashboardButton = '';
-    if (this.props.isUserLoggedIn) {
-      dashboardButton = (
-        <a
-          className="btn btn-primary"
-          href="#"
-          onClick={() => this.props.changePage('dashboard')}
-          >
-          dashboard
-        </a>
-      );
-    }
-    return dashboardButton;
-  }
-
-  getLoginButton() {
-    let loginButton = '';
-    if (!this.props.isUserLoggedIn) {
-      loginButton = (
-        <a
-          className="btn btn-primary"
-          href="#"
-          onClick={() => this.onScreenShow()}
-          >
-          Sign In
-        </a>
-      );
-    }
-    return loginButton;
-  }
-
-  onScreenShow() {
-    document.getElementById('login').style.display = 'block';
   }
 }
